refactor: unify duplicated diff tree builders and async queue logic
Extract shared useDiffQueue hook to eliminate duplicated concurrency
queue implementations between useMultiFileDiff and useBranchDiff.
Extract shared diffTree utility (buildWorkingTree, buildBranchTree)
with a common ensurePath helper to replace duplicate tree-building
functions in DiffFileList.

Closes #29
